DRY up the dashboard views a tad
As pointed out by sameerav in
https://github.com/calblueprint/revolv/pull/223#discussion_r28392817,
the code for the Admin and Ambassador dashboard views class had a lot of
repeated code in it. Added in a BaseStaffDashboardView class to abstract
the common functionality.

Note that I found the DonorDashboardView different enough to warrant not
having it inherit from BaseStaffDashboardView, which is why I called it
BaseStaffDashboardView rather than BaseDashboardView.