tweak final dashboard styles a bit
As suggested by andylouisqin in
https://github.com/calblueprint/revolv/pull/223#issuecomment-93247356,
the line height for the text in the project description on the
dashboards was a little too short, making it a bit hard to read. Upped
it to 22px (24px was suggested, but that seems VERY large to me).

Also made the links that select roles on the dashboard non underlines by
default and underline on hover instead.