Tune up dashboard create project test
It now works! I had to do a couple of things. Basically, the util which
actually fills the form on the project create/edit page just makes a new
modelform and fills it with the fields of the project in question. This
actually works out of the box, even though the project doesn't have a
cover_photo or categories. I had thought that categories were required
before, but I guess they're not.

I still put in some custom form_filler classes that are not really used
by the test class right now, but if they don't need to be used by any
other tests at the end of this branch I'll take them out.