(#290) Serve font awesome ourselves
Turns out that the UC Berkeley proxy would sometimes try to block the
loading of font awesome through MaxCDN, so I figured that to avoid
issues like this (which would cause the dashboard role select chevron to
not appear), we should just server font awesome ourselves.

I added it to bower.json and added a foundation-settings file to serve
the fonts from the right location also. It serves it correctly from
dashboard.css now, and loading times won't be very much longer because
font awesome is tiny compared to foundation, which we're already loading
inline. Anyway, this fixes #290.