Dashboard is fully tabletified
Had to do some tricky, kind of cool stuff. On tablet landscape, it's
pretty much the same except some things are smaller and/or moved around
(particularly, the "(ACTIVE)" project status container is on a new
line). On tablet portrait, there's not enough room, so the sidebar
instead floats to the left overlaying the project container, which is
full width. You can then open and close the sidebar with a slick
animation.

Still to come is making this work for mobile, but should work in pretty
much the same way.