Show first project, don't close sidebar on large
Two things for this commit:

In order to make sure the user can never get stuck on mobile by
closing the sidebar when there is no project selected on the dashboard,
I added in some javascript to always select the first project, if there
is one. If there are no projects, they're out of luck, but this will
never be the case.

Previously, it had been that the first project only got displayed by
default if it was part of the first project group - this was a bug which
is now fixed.

Also, I added in some javascript to detect the size of the screen when
clicking on buttons in the sidebar - if on tablet portrait or smaller (<
800px width) then the sidebar will automatically close, but otherwise it
will not close.

Getting closer to having everything fully mobilified.