Always show dash sidebar on switch to large screen
The dashboard sidebar is supposed to float (and be able to be hidden) on
tablet portrait orientation and NOT on landscape orientation (on
landscape, it should be fixed to the left, and not collapsable). That
used to mean that if you turned the tablet to portrait orientation,
closed the sidebar, and turned it back to landscape orientation, you
would never be able to get the sidebar back, because there's no sidebar
button on landscape orientation.

I fixed that by adding some javascript which will autoshow the sidebar
if the window is ever resized to be greater than tablet mobile width
(which is defined to be 800px, for now). We can change that breakpoint
in the js if we need to later.