Add in RevolvUserProfileSeedSpec
This will be the new way to seed data for RevolvUserProfile. Going to
test it presently. Also, set up a bit of infrastructure for defining how
seeding works: basically, you can define a SeedSpec subclass which has
two methods: seed() and clear(). Hopefully this will be a nice way to
impelement different kinds of seeds.