Payment seed spec
Had to fix another bug too: it turned out that in project/factories.py,
we were always creating a completed project to use for a repayment
factory on import of the project/factories.py module. This created a
dummy user, user_0 when the module was imported.

This was, of course, fine in tests, but since we use factories in the
seed script, it happened in the seed script too. Thus, the first time
the seed script ran, it worked fine, because it produced user_0.
However, if you tried to ever clear then run it again, it would complain
about user_0 being a duplicated username.

So, I made that completed project in the factories module a
LazyAttribute instead, and now it works.