Crude styles for custom pages and title updates
Custom pages have some semblence of style now. Images don't seem to be
resizing correctly: this may be because of local settings and may work
on staging? Documents seem to be not working correctly also.

Also, I went through and changed all the titles of the pages to match
<title> | RE-volv format. There are a bunch of files which I think
should not even be here...gunna delete them, I think.

Also, need to make sure the administrators can log in and see the
wagtail admin, because I think only superusers can now.