Finish CMS seed spec and a test for it
Some tricky stuff here actually: because wagtail's Page model is part of
a django-treebeard tree, you can't simply .all().delete() the pages to
clear seed data because there is meta data (paths, or something like
that, I think) that is stored in the parent node. In order to delete the
page models, we actually have to delete the root Page of the Site and
make a new one.

Still need to document the seed spec, will do that shortly.