Fix top bar colors
For the longest time, all the topbar link colors have been darker than
they should have been. This was because we kind of screwed up the
foundation settings for the top bar: we were trying to override
foundation settings in foundation-settings.scss, but we then overrode
THOSE settings in nav.scss. This commit fixes those shenannigans.