(#339) Fix nav bar styles to accomodate many pages
In order to do this, I had to make the nav bar visible (i.e. not
hamburger menu'd in) for only large screens, not tablets. This is
necessary because there are so many nav items on RE-volv's site right
now that if we wanted to display them on portrait tablet, the text would
have to be so small that it would be unreadable.

Anyway, going to do the footer next.