Fix styles for CMSified footer bar
Made it so that there are maximum of four sections per line in the
footer bar, and on small screens they expand to take the entire width.
Also added the 100 weight for Lato and fixed a bunch of font-weights
which haven't worked since we switched from Helvetica Neue to Lato.