Update settings to serve staticfiles on s3
Wasn't too big of a change, actually. Had to make sure that the bucket
settings are correct (this is something that we will need to put in the
documentation), and it only took a few settings updates, because this
functionality is built in with boto and django-storages.

Had to define a few custom storages to make sure staging and production
staticfiles go in different places, but I did my best to separate and
document the changes.