Fixes #366: Add settings to db seed
So, before this, there were wagtail settings models defined, with
default values, but in order to get them to actually show up on pages,
you had to go to the settings page and save each one.

This is because the register_setting decorator doesn't actually create
instances of the models (nor should it). So, to eliminate unecessary
work, I added the cms models to a seed file. It also automatically seeds
all registered settings, so no need to worry about editing the seed spec
for settings after defining a new setting.