Add donor table styles, fix timedelta bug
Added the styles for the donor table (#321). Also, fixed a super weird
bug that I think has been in the codebase for like a year:
Project.days_until_end was returning a negative number if the end date
of the project was in the future. Which was no bueno. So I fixed that an
added a test for it.