Tweak settings files to work better.
STATICFILES_FINDERS wasn't copied from original and needs to be there,
at least for some tests. Error was made in omitting BASE_DIR from
EMAIL_TEMPLATE_PATH, so restor that. Revert a couple of _SECURE
cookie settings to False on dev since requiring them to be True on dev
leads to problems as we don't have a secure dev server.