Changing top nav buttons on home page
Per the client's request, this changes the appearance of the home page's top nav element so that:

1. When the user is scrolled above the bottom point of the hero video, the nav shows side-by-side buttons saying 'Sign Up' and 'Sign In' (or 'Dashboard' and 'Sign Out' if signed in).
2. When the user is scrolled down below aforementioned bottom point, the nav shows the 'Donate' button.

This does not happen on other pages, as it is not appropriate there (the "always see the donate button" imperative looks different there).

The nav still looks awful on small screen, but that is not new to this branch.