Removing flex layout from body
Indiscriminately using `display: flex` on the `body` element was causing layout issues in non-content pages.

This leaves the flexbox layout in `cms-page.scss` but removes it from `body` elsewhere. The result is a footer that doesn't stick to the bottom of the page, but also an elimination of those pesky layout errors.