Reimplementing h1 header on Create Project page
The weird thing where the `h1` element with the page's title was supposed to be floated over the header … let's not do that. Instead this simply places the `h1` inside the header.

This is accomplished by creating a new override block inside the `nav.html` template (or rather, `project-nav.html`, which extends it) that lets you inject stuff into the header as you wish.