Adding Toastr CSS link
The CDN link for the Toastr library's styles was never added to the project. The predictable bad result was the weird, bad stuff the Toastr plugin did when actually trying to display messages.

Why was the CSS not linked in? Probably this is one of those situations where the Toastr code was partway through being incorporated when the project passed over to new developers' hands.

The styles are not particularly consistent with the rest of the site's styling. *But this is a small problem compared to there being no styles at all*, so this patch is an improvement.