Adding custom testimonials
The hard-coded placeholder testimonials are replaced by (unfortunately hard-coded) real testimonials.

Images lifted from the Google Doc are added alongside appropriate text and captions.

The styles of testimonials on mobile devices are changed. The float-based layout was breaking in a bizarre way. This arranges the images in the desired "inline" fashion without similar breakage (although the result is still kinda funny-looking).