Changing "Dashboard" to "My Portfolio"
This makes some superficial changes to turn "Dashboard" into "My Portfolio".

For the most part, the code still looks as though this feature were called "Dashboard". But we *really* cannot afford to go through and comb through the 300+ references to the "Dashboard" before the launch date. Hence the superficiality of this fix.