New placeholder image
This adds a new, better placeholder image (= one without a stock photo watermark!) to the home page and project list page.

It also makes that image's element more robust by getting rid of the `img` and replacing it with a `.img` box that uses the image as its background image. This makes it possible to use the `background-size: cover` CSS rule to ensure that the image, no matter what its dimensions are, is always stretched in such a way as to ensure that the image is centered and covers the visible area.